+++ libs/components/src/lib/components/table/index.ts
@@ -19,7 +19,7 @@ export * from './td/td.component';
 export * from './th/th.component';
 export * from './directives/tree-loading.directive';
 export * from './th-group/th-group.component';
-export * from './tr/tr.component';
+export * from './tr';
 export * from './data-source';
 export * from './table.types';
 export * from './table.const';