+++ libs/components/src/lib/components/table/td/td.service.ts
@@ -1,0 +1,19 @@
+import { Injectable } from '@angular/core';
+import { BehaviorSubject } from 'rxjs';
+
+@Injectable()
+export class PrizmTdService {
+  private readonly count$$ = new BehaviorSubject<number>(0);
+  public readonly count$ = this.count$$.asObservable();
+  get count() {
+    return this.count$$.value;
+  }
+
+  public increment(value = 1) {
+    this.count$$.next(this.count$$.value + value);
+  }
+
+  public decrement(value = 1) {
+    this.count$$.next(this.count$$.value - value);
+  }
+}